Refactor AI chat components and integrate new LLM streaming logic
- Updated `AiChatWindow` to utilize the new `useAiChat` hook for improved state management and role handling.
- Introduced `ChatView` component to separate chat UI logic, enhancing code organization and readability.
- Implemented `streamLlm` function for robust LLM streaming with fallback model support, ensuring reliability in AI responses.
- Enhanced prompt templates for AI roles, improving contextual understanding and response accuracy.
- Added client identification logic to track user interactions more effectively.
