+++ src/components/AiChatWindow.tsx
@@ -1,20 +1,8 @@
 'use client';
 
-import React, { useState, useRef, useEffect, memo } from 'react';
-import { useChat } from 'ai/react';
-import { BookOpen, Sparkles, X, MoreHorizontal, Send, RefreshCw, AlertTriangle, FileText, Eraser, Loader } from 'lucide-react';
-import ReactMarkdown from 'react-markdown';
-import remarkMath from 'remark-math';
-import rehypeKatex from 'rehype-katex';
-import 'katex/dist/katex.min.css';
-import { Prism as SyntaxHighlighter } from 'react-syntax-highlighter';
-import { oneDark } from 'react-syntax-highlighter/dist/esm/styles/prism';
-import { useChatStore } from '@/store/chat';
-import type { ChatStoreState, ChatMessage } from '@/store/chat';
-import { Message } from 'ai';
-
-// 定義 AI 角色的類型
-export type AiRole = 'specialist' | 'guide';
+import React, { useState, useEffect } from 'react';
+import { useAiChat, type AiRole } from '@/hooks/useAiChat';
+import ChatView from './ChatView';
 
 // 定義元件的 Props
 interface AiChatWindowProps {
@@ -29,33 +17,6 @@ interface AiChatWindowProps {
   startWithSummary?: boolean;
 }
 
-// 用於管理不同角色 UI 和行為的設定物件
-const roleConfig = {
-  specialist: {
-    icon: <BookOpen className="h-6 w-6 text-indigo-500" />,
-    title: '文章 AI 助理',
-    welcomeMessage: '嘿！對這篇文章有什麼好奇的嗎？無論是想快速抓重點，還是深入探討，隨時都能問我！',
-    placeholder: '針對這篇文章提問...'
-  },
-  guide: {
-    icon: <Sparkles className="h-6 w-6 text-amber-500" />,
-    title: '全站數位分身',
-    welcomeMessage: `您好，歡迎來到 Code Lab！我是 Peienwu 的 AI 數位分身 **Bryan**，一個由 Gemini AI 驅動的智慧嚮導。我已經學習了這個網站中所有的技術文章、專案與個人背景。
-
-無論您想尋找特定資訊、理解複雜觀念，或對 Peienwu 本人感到好奇，我都樂意提供協助。
-
-**您可以試著這樣問我：**
-- 「Peienwu 是誰？他做過哪些專案？」
-- 「這個網站的 AI 數位分身是怎麼實現的？」
-
-請隨時輸入您的問題，我很樂意為您指引！`,
-    placeholder: '與 Peien 的數位分身對話...'
-  }
-};
-
-const SUMMARY_PROMPT = '請為我生成這篇文章的摘要';
-
-// 移除 forwardRef
 export default function AiChatWindow({ 
   initialRole, 
   chatKey,
@@ -67,128 +28,41 @@ export default function AiChatWindow({
   allowRoleSwitching = true,
   startWithSummary = false
 }: AiChatWindowProps) {
-  // get stored messages
-  const getMessages = useChatStore((state: ChatStoreState) => state.getMessages);
-  const setMessagesInStore = useChatStore((state: ChatStoreState) => state.setMessages);
-  const clearSession = useChatStore((state: ChatStoreState) => state.clearSession);
-
-  const stored = getMessages(chatKey);
-
-  const defaultWelcome: Message[] = [{ id: 'initial-welcome', role: 'assistant', content: roleConfig[initialRole].welcomeMessage }];
-
-  const [currentRole, setCurrentRole] = useState<AiRole>(initialRole);
-  
-  // -- 新增狀態 --
+  // UI state
   const [isSwitchMenuOpen, setSwitchMenuOpen] = useState(false);
   const [isConfirmModalOpen, setConfirmModalOpen] = useState(false);
   const [isClearConfirmOpen, setClearConfirmOpen] = useState(false);
   const [roleToSwitch, setRoleToSwitch] = useState<AiRole | null>(null);
-  const switchMenuRef = useRef<HTMLDivElement>(null);
-  const chatContainerRef = useRef<HTMLDivElement>(null);
-  
-  // -- 新增動畫狀態 --
   const [isVisible, setIsVisible] = useState(false);
 
-  // 在每次請求期間避免重複顯示錯誤訊息
-  const errorShownRef = useRef(false);
+  // Chat logic from hook
+  const chatState = useAiChat({
+    initialRole,
+    chatKey,
+    initialInputValue,
+    onUnmount,
+    articleContent,
+    startWithSummary,
+  });
 
+  // Animation effect
   useEffect(() => {
-    // 元件掛載後觸發進場動畫
     const timer = setTimeout(() => setIsVisible(true), 10);
     return () => clearTimeout(timer);
   }, []);
   
+  // Handlers
   const handleClose = () => {
-    setIsVisible(false); // 觸發退場動畫
+    setIsVisible(false);
     setTimeout(() => {
-      onClose(); // 動畫結束後才真正關閉
-    }, 300); // 需與 CSS transition duration 一致
+      onClose();
+    }, 300);
   };
 
-  const { icon, title, placeholder } = roleConfig[currentRole];
-  
-  // -- 整合 useChat --
-  const { messages, input, handleInputChange, handleSubmit, setMessages, append, setInput } = useChat({
-    api: currentRole === 'specialist' ? '/api/article-chat' : '/api/chat',
-    initialMessages: stored.length ? stored : defaultWelcome,
-    body: { articleContent: currentRole === 'specialist' ? articleContent : undefined },
-    onError(error) {
-      if (!errorShownRef.current) {
-        append({
-          id: `error-${Date.now()}`,
-          role: 'assistant',
-          content: '抱歉，AI 目前暫時無法回應（可能是額度用盡或連線問題）。請稍後再試一次！',
-        });
-        errorShownRef.current = true;
-      }
-      console.error('[AiChatWindow] AI error:', error);
-    },
-  });
-
-  const isLoading = messages.length > 0 && messages[messages.length - 1].role === 'user';
-
-  const inputRef = useRef(input);
-  useEffect(() => {
-    inputRef.current = input;
-  }, [input]);
-
-  // 在元件掛載時設定初始輸入值
-  useEffect(() => {
-    if (initialInputValue) {
-      setInput(initialInputValue);
-    }
-  }, []);
-
-  // 在元件卸載時保存當前輸入值
-  useEffect(() => {
-    return () => {
-      onUnmount(inputRef.current);
-    };
-  }, [onUnmount]);
-
-  // persist whenever messages changes
-  useEffect(() => {
-    const messagesToStore = messages.filter(
-      (m) => m.role === 'user' || m.role === 'assistant'
-    );
-    setMessagesInStore(chatKey, messagesToStore);
-  }, [messages]);
-
-  // 自動捲動至最新訊息
-  useEffect(() => {
-    if (!chatContainerRef.current) return;
-
-    const isAtBottom = chatContainerRef.current.scrollHeight - chatContainerRef.current.scrollTop <= chatContainerRef.current.clientHeight + 1; // +1 for tolerance
-    const lastMessage = messages[messages.length - 1];
-    
-    // 條件 1: AI 剛開始回覆 (前一則是 user，現在是 assistant)
-    const isAiStartingToReply = messages.length > 1 && messages[messages.length - 2].role === 'user' && lastMessage.role === 'assistant';
-
-    if (isAiStartingToReply) {
-      const userMessageId = `message-${messages[messages.length - 2].id}`;
-      const userMessageElement = chatContainerRef.current.querySelector(`#${userMessageId}`);
-      if (userMessageElement) {
-        userMessageElement.scrollIntoView({ behavior: 'smooth', block: 'start' });
-        return; // 捲動到指定位置後就停止
-      }
-    }
-
-    // 條件 2: 使用者自己捲動到底部時，或自己發出訊息時，保持在底部
-    if (isAtBottom || lastMessage.role === 'user') {
-      chatContainerRef.current.scrollTop = chatContainerRef.current.scrollHeight;
-    }
-  }, [messages]);
-
-  // 自動觸發摘要
-  useEffect(() => {
-    if (startWithSummary && currentRole === 'specialist' && messages.length === 1) {
-      append({ role: 'user', content: SUMMARY_PROMPT });
-    }
-  }, [startWithSummary, currentRole]);
-
-  // 按鈕觸發摘要
-  const handleGenerateSummaryClick = () => {
-    append({ role: 'user', content: SUMMARY_PROMPT });
+  const handleSwitchRole = (role: AiRole) => {
+    setRoleToSwitch(role);
+    setConfirmModalOpen(true);
+    setSwitchMenuOpen(false);
   };
 
   const confirmSwitch = () => {
@@ -197,284 +71,47 @@ export default function AiChatWindow({
     }
     setConfirmModalOpen(false);
   };
-  
-  const handleClearChat = () => {
-    clearSession(chatKey);
-    setMessages(defaultWelcome);
+
+  const handleClearChatWithModal = () => {
+    chatState.handleClearChat();
     setClearConfirmOpen(false);
   };
 
-  // 點擊外部關閉切換菜單
-  useEffect(() => {
-    function handleClickOutside(event: MouseEvent) {
-      if (switchMenuRef.current && !switchMenuRef.current.contains(event.target as Node)) {
-        setSwitchMenuOpen(false);
-      }
-    }
-    document.addEventListener("mousedown", handleClickOutside);
-    return () => {
-      document.removeEventListener("mousedown", handleClickOutside);
-    };
-  }, [switchMenuRef]);
-
-  const handleSwitchRole = (role: AiRole) => {
-    setRoleToSwitch(role);
-    setConfirmModalOpen(true);
-    setSwitchMenuOpen(false);
-  };
-
-  const otherRole = currentRole === 'specialist' ? 'guide' : 'specialist';
-  const showSummaryButton = currentRole === 'specialist' && messages.length === 1;
-
-  // --- 程式碼高亮自訂元件 ---
-  const CodeBlock = memo(({ node, inline, className, children, ...props }: any) => {
-    const codeString = String(children).replace(/\n$/, '');
-
-    // 處理行內程式碼
-    if (inline || (!codeString.includes('\n') && !(className || '').includes('language-'))) {
-      return (
-        <code className="bg-slate-100 dark:bg-slate-800 rounded px-1.5 py-0.5 font-mono text-sm text-pink-600 dark:text-pink-400 whitespace-pre">
-          {codeString}
-        </code>
-      );
-    }
-
-    // 處理區塊程式碼
-    const match = /language-(\w+)/.exec(className || '');
-    const lang = match ? match[1] : undefined;
-
-    return (
-      <SyntaxHighlighter
-        language={lang}
-        style={oneDark}
-        // 模擬 rehype-pretty-code 的行為，只用一個 pre 標籤包裹
-        PreTag={(preProps) => <pre className="not-prose" {...preProps} />}
-        // 確保 highlighter 本身沒有多餘的樣式
-        customStyle={{
-          background: 'transparent',
-          margin: 0,
-          padding: 0,
-          border: 'none',
-        }}
-        codeTagProps={{
-          style: {
-            fontFamily: 'inherit',
-            fontSize: 'inherit',
-            lineHeight: 'inherit',
-            padding: '1rem',
-            display: 'block',
-            width: '100%',
-            overflowX: 'auto',
-          }
-        }}
-        {...props}
-      >
-        {codeString}
-      </SyntaxHighlighter>
-    );
-  });
-
-  // 封裝 submit，重置錯誤旗標
-  const onFormSubmit = (e: React.FormEvent<HTMLFormElement>) => {
-    errorShownRef.current = false;
-    handleSubmit(e);
-  };
-
   return (
-    <>
-      {/* 清除確認 Modal */}
-      {isClearConfirmOpen && (
-        <div className="fixed inset-0 bg-black bg-opacity-50 z-[60] flex items-center justify-center">
-          <div className="bg-white dark:bg-slate-800 rounded-lg shadow-xl p-6 m-4 max-w-sm text-center">
-            <AlertTriangle className="h-12 w-12 text-red-500 mx-auto mb-4" />
-            <h3 className="text-lg font-semibold text-slate-900 dark:text-slate-100 mb-2">清除對話紀錄</h3>
-            <p className="text-sm text-slate-600 dark:text-slate-400 mb-6">
-              您確定要清除目前的對話紀錄嗎？這個操作無法復原。
-            </p>
-            <div className="flex justify-center gap-4">
-              <button 
-                onClick={() => setClearConfirmOpen(false)}
-                className="px-4 py-2 rounded-md text-sm font-medium bg-slate-200 dark:bg-slate-700 text-slate-800 dark:text-slate-200 hover:bg-slate-300 dark:hover:bg-slate-600">
-                取消
-              </button>
-              <button 
-                onClick={handleClearChat}
-                className="px-4 py-2 rounded-md text-sm font-medium bg-red-600 text-white hover:bg-red-700">
-                確認清除
-              </button>
-            </div>
-          </div>
-        </div>
-      )}
-
-      {/* 切換確認 Modal */}
-      {isConfirmModalOpen && (
-        <div className="fixed inset-0 bg-black bg-opacity-50 z-[60] flex items-center justify-center">
-          <div className="bg-white dark:bg-slate-800 rounded-lg shadow-xl p-6 m-4 max-w-sm text-center">
-            <AlertTriangle className="h-12 w-12 text-amber-500 mx-auto mb-4" />
-            <h3 className="text-lg font-semibold text-slate-900 dark:text-slate-100 mb-2">切換 AI 角色</h3>
-            <p className="text-sm text-slate-600 dark:text-slate-400 mb-6">
-              我們將為新角色開啟全新的對話；先前對話會保留，稍後仍可返回瀏覽。
-            </p>
-            <div className="flex justify-center gap-4">
-              <button 
-                onClick={() => setConfirmModalOpen(false)}
-                className="px-4 py-2 rounded-md text-sm font-medium bg-slate-200 dark:bg-slate-700 text-slate-800 dark:text-slate-200 hover:bg-slate-300 dark:hover:bg-slate-600">
-                取消
-              </button>
-              <button 
-                onClick={confirmSwitch}
-                className="px-4 py-2 rounded-md text-sm font-medium bg-blue-600 text-white hover:bg-blue-700">
-                確認切換
-              </button>
-            </div>
-          </div>
-        </div>
-      )}
-
-      {/* -- Use Gentle Fade-in & Slide-up Animation -- */}
-      <div className={`fixed inset-x-0 top-16 bottom-0 sm:top-auto sm:inset-x-auto sm:right-4 sm:bottom-4 z-50 w-full sm:max-w-md transform-gpu transition-all duration-300 ease-in-out ${isVisible ? 'opacity-100 translate-y-0' : 'opacity-0 translate-y-10'}`}>
-        <div className="bg-white dark:bg-slate-900 rounded-t-lg sm:rounded-lg shadow-2xl flex flex-col h-full sm:h-[80vh] sm:max-h-[800px] sm:border border-slate-200 dark:border-slate-700">
-          {/* 標頭 */}
-          <header className="flex items-center justify-between border-b border-slate-200 dark:border-slate-700 p-3">
-            <div className="flex items-center gap-3">
-              {icon}
-              <h2 className="font-semibold text-slate-800 dark:text-slate-200">{title}</h2>
-            </div>
-            <div className="flex items-center gap-1 relative">
-              {/* --- 條件渲染切換按鈕 --- */}
-              {allowRoleSwitching && (
-                <div ref={switchMenuRef}>
-                  <button 
-                    onClick={() => setSwitchMenuOpen(!isSwitchMenuOpen)}
-                    className="p-2 rounded-full text-slate-500 hover:bg-slate-100 dark:text-slate-400 dark:hover:bg-slate-800"
-                  >
-                    <MoreHorizontal className="h-5 w-5" />
-                  </button>
-                  {isSwitchMenuOpen && (
-                    <div className="absolute top-full right-0 mt-2 w-56 bg-white dark:bg-slate-800 rounded-md shadow-lg border border-slate-200 dark:border-slate-700 z-10">
-                      <button
-                        onClick={() => handleSwitchRole(otherRole)}
-                        className="flex items-center gap-3 w-full text-left px-4 py-3 hover:bg-slate-50 dark:hover:bg-slate-700 transition-colors"
-                      >
-                        <RefreshCw className="h-5 w-5 text-slate-500 dark:text-slate-400" />
-                        <div className="text-slate-800 dark:text-slate-200">
-                          切換到 <span className="font-semibold">{roleConfig[otherRole].title}</span>
-                        </div>
-                      </button>
-                    </div>
-                  )}
-                </div>
-              )}
-              <button
-                onClick={() => setClearConfirmOpen(true)}
-                className="p-2 rounded-full text-slate-500 hover:bg-slate-100 dark:text-slate-400 dark:hover:bg-slate-800"
-                aria-label="清除對話紀錄"
-              >
-                <Eraser className="h-5 w-5" />
-              </button>
-              <button
-                onClick={handleClose}
-                className="p-2 rounded-full text-slate-500 hover:bg-slate-100 dark:text-slate-400 dark:hover:bg-slate-800"
-                aria-label="關閉聊天視窗"
-              >
-                <X className="h-5 w-5" />
-              </button>
-            </div>
-          </header>
-
-          {/* 對話區域 */}
-          <div ref={chatContainerRef} className="flex-1 overflow-y-auto p-4 flex flex-col">
-            {messages.map((msg, index) => {
-              const isUser = msg.role === 'user';
-              const isLastInSequence = !messages[index + 1] || messages[index + 1].role !== msg.role;
-              const isFirstInSequence = !messages[index - 1] || messages[index - 1].role !== msg.role;
-
-              return (
-                <div 
-                  key={msg.id} 
-                  id={`message-${msg.id}`}
-                  className={`flex items-end gap-2.5 ${isUser ? 'flex-row-reverse' : ''} ${isFirstInSequence ? 'mt-4' : 'mt-1'}`}
-                >
-                  {msg.role === 'assistant' && (
-                    <div className="flex-shrink-0">
-                      {isLastInSequence ? (
-                         currentRole === 'guide' ? (
-                          <img src="/images/avatar.jpeg" alt="Peien Wu" className="h-8 w-8 rounded-full" />
-                        ) : (
-                          <div className="h-8 w-8 rounded-full bg-indigo-100 dark:bg-indigo-900 flex items-center justify-center ring-1 ring-inset ring-indigo-200 dark:ring-indigo-800">
-                            <BookOpen className="h-5 w-5 text-indigo-500" />
-                          </div>
-                        )
-                      ) : (
-                        <div className="w-8"></div>
-                      )}
-                    </div>
-                  )}
-                  <div className={`px-4 py-2 rounded-2xl max-w-[85%] ${isUser ? 'bg-blue-500 text-white' : 'bg-slate-100 dark:bg-slate-700'} ${isLastInSequence ? (isUser ? 'rounded-br-none' : 'rounded-bl-none') : ''}`}>
-                    <div className={`prose prose-sm max-w-full ${isUser ? 'text-white' : 'dark:prose-invert'}`}>
-                      {msg.role === 'assistant' ? (
-                        <ReactMarkdown 
-                          remarkPlugins={[remarkMath] as any}
-                          rehypePlugins={[rehypeKatex] as any}
-                          components={{
-                            code: CodeBlock,
-                          }}
-                        >
-                          {msg.content}
-                        </ReactMarkdown>
-                      ) : (
-                        <p>{msg.content}</p>
-                      )}
-                    </div>
-                  </div>
-                </div>
-              )
-            })}
-            {/* -- 新增摘要按鈕 -- */}
-            {showSummaryButton && (
-              <div className="flex justify-center my-4">
-                <button
-                  onClick={handleGenerateSummaryClick}
-                  className="inline-flex items-center gap-2 px-3 py-2 text-sm font-medium text-blue-600 bg-blue-50 dark:bg-blue-900/50 dark:text-blue-400 rounded-lg hover:bg-blue-100 dark:hover:bg-blue-900 transition-colors"
-                >
-                  <FileText className="h-4 w-4" />
-                  一鍵生成本文摘要
-                </button>
-              </div>
-            )}
-
-            {/* --- Loading Indicator --- */}
-            {isLoading && (
-              <div className="flex items-center justify-center p-4">
-                <Loader className="h-6 w-6 text-slate-400 animate-spin" />
-                <p className="ml-2 text-sm text-slate-500">處理中...</p>
-              </div>
-            )}
-          </div>
-
-          {/* 輸入框 */}
-          <div className="border-t border-slate-200 dark:border-slate-700 p-3 bg-white dark:bg-slate-900">
-            <form onSubmit={onFormSubmit} className="flex items-center gap-2">
-              <input
-                type="text"
-                value={input}
-                onChange={handleInputChange}
-                placeholder={placeholder}
-                className="w-full px-4 py-2 bg-slate-100 dark:bg-slate-800 border border-transparent rounded-full focus:outline-none focus:ring-2 focus:ring-blue-500"
-              />
-              <button
-                type="submit"
-                className="p-3 bg-blue-600 text-white rounded-full hover:bg-blue-700 disabled:bg-blue-300 disabled:cursor-not-allowed"
-                disabled={!input.trim()}
-                aria-label="傳送訊息"
-              >
-                <Send className="h-5 w-5" />
-              </button>
-            </form>
-          </div>
-        </div>
-      </div>
-    </>
+    <ChatView
+      // State from hook
+      messages={chatState.messages}
+      input={chatState.input}
+      currentRole={chatState.currentRole}
+      isLoading={chatState.isLoading}
+      showSummaryButton={chatState.showSummaryButton}
+      otherRole={chatState.otherRole}
+      
+      // UI state
+      isVisible={isVisible}
+      isSwitchMenuOpen={isSwitchMenuOpen}
+      isConfirmModalOpen={isConfirmModalOpen}
+      isClearConfirmOpen={isClearConfirmOpen}
+      
+      // Handlers from hook
+      handleInputChange={chatState.handleInputChange}
+      handleFormSubmit={chatState.handleFormSubmit}
+      handleGenerateSummary={chatState.handleGenerateSummary}
+      handleClearChat={handleClearChatWithModal}
+      
+      // UI handlers
+      handleClose={handleClose}
+      handleSwitchRole={handleSwitchRole}
+      setSwitchMenuOpen={setSwitchMenuOpen}
+      setConfirmModalOpen={setConfirmModalOpen}
+      setClearConfirmOpen={setClearConfirmOpen}
+      confirmSwitch={confirmSwitch}
+      
+      // Options
+      allowRoleSwitching={allowRoleSwitching}
+    />
   );
-} 
+}
+
+// Re-export the type for backward compatibility
+export type { AiRole };
